доделал с JdbcUserMealRepositoryImpl

+++ src/main/java/ru/javawebinar/topjava/repository/jdbc/JdbcUserMealRepositoryImpl.java
@@ -4,6 +4,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.BeanPropertyRowMapper;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
+import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
 import ru.javawebinar.topjava.model.User;
@@ -11,6 +12,7 @@ import ru.javawebinar.topjava.model.UserMeal;
 import ru.javawebinar.topjava.repository.UserMealRepository;
 
 import javax.sql.DataSource;
+import java.sql.Timestamp;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -26,6 +28,9 @@ public class JdbcUserMealRepositoryImpl implements UserMealRepository {
     @Autowired
     private JdbcTemplate jdbcTemplate;
 
+    @Autowired
+    private NamedParameterJdbcTemplate namedParameterJdbcTemplate;
+
     private SimpleJdbcInsert insertMeal;
 
     public JdbcUserMealRepositoryImpl(DataSource dataSource) {
@@ -38,32 +43,45 @@ public class JdbcUserMealRepositoryImpl implements UserMealRepository {
         MapSqlParameterSource map = new MapSqlParameterSource()
                 .addValue("time",userMeal.getDateTime())
                 .addValue("calories",userMeal.getCalories())
-                .addValue("description",userMeal.getDescription());
+                .addValue("description",userMeal.getDescription())
+                .addValue("user_id",userId);
+        if (userMeal.isNew()) {
+            Number newKey = insertMeal.executeAndReturnKey(map);
+            userMeal.setId(newKey.intValue());
+        } else {
+            namedParameterJdbcTemplate.update(
+                    "UPDATE meals SET time=: time, calories=: calories, description=: description " +
+            "WHERE meal_id=:meal_id AND user_id=: user_id",map);
+        }
         return userMeal;
     }
 
     @Override
     public boolean delete(int id, int userId) {
-        return jdbcTemplate.update("DELETE FROM meals WHERE meal_id=?",id) !=0;
+        return jdbcTemplate.update("DELETE FROM meals WHERE meal_id=? AND user_id=?", id, userId) !=0;
     }
 
     @Override
     public UserMeal get(int id, int userId) {
-        return jdbcTemplate.queryForObject("SELECT * FROM meals WHERE meal_id=?",ROW_MAPPER, id);
+        return jdbcTemplate.queryForObject("SELECT * FROM meals WHERE meal_id=? AND user_id=?",ROW_MAPPER, id, userId);
     }
 
     @Override
     public List<UserMeal> getAll(int userId) {
-        return null;
+        return jdbcTemplate.query("SELECT * FROM meals ORDER BY user_id", ROW_MAPPER);
     }
 
     @Override
     public void deleteAll(int userId) {
+        jdbcTemplate.update("DELETE FROM meals WHERE user_id=?", userId);
 
     }
 
     @Override
     public List<UserMeal> getBetween(LocalDateTime startDate, LocalDateTime endDate, int userId) {
-        return null;
+        Timestamp startD = Timestamp.valueOf(startDate);
+        Timestamp endD = Timestamp.valueOf(endDate);
+        return jdbcTemplate.query(
+                "SELECT * FROM meals WHERE user_id=? AND time BETWEEN ? AND ?",ROW_MAPPER, userId,startD, endD);
     }
 }
